requirements.txt
git+https://github.com/Rapptz/discord.py@master
psutil
timeago==1.0.13
expiringdict
pixivpy